Fixes #39. Fixes tabcontainer bug introduced in last update. Closes #40.
Refactors how thread data is passed between functions and cleans up how switching between threads works. Rewrites refreshing to properly call downloadThread. Fixes parameters not being passed when waiting for the document to load to insert tabs. Fixes trying to apply the new layout class to the allowOnChannelContainer when it did not exist.